(wip) consolidating params
